created raw conent component
